requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
radon>=6.0.0          # Complexity metrics
bandit>=1.7.0         # Security scanning
rich>=13.0.0          # Pretty console output

# Web & Database
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.6
psutil>=5.9.0
gputil>=1.4.0

# Optional but recommended
ruff>=0.1.0          # Fast Python linter with auto-fix
pytest>=7.0.0        # Testing framework

# For development
black>=23.0.0        # Code formatting
mypy>=1.0.0          # Type checking
